+++ src/photos.h
@@ -64,7 +64,7 @@ protected slots:
     void createFolder();
     void commandFinished(int cmd, bool error, const FileInfoList &results);
     void filesDropped(const QList<QUrl> &files, const QUrl &destination);
-    void putProgress(qint64 done, qint64 total);
+    void putProgress(int done, int total);
 
 private:
     bool busy;
